In startUpConverter(), use 'int32_t' rather than 'long' for the IQ array, to match the change in the upconverter API.
git-svn-id: http://svn.eol.ucar.edu/svn/ka/trunk@87 3b647ff1-7a5c-455a-93f8-4e24338e3c8e

+++ src/kadrx/main.cpp
@@ -217,7 +217,7 @@ void startUpConverter(Pentek::p7142Up& upConverter,
 
 	// create the signal
 	unsigned int n = pulsewidth_counts * 2;
-	long IQ[n];
+	int32_t IQ[n];
 
 	for (unsigned int i = 0; i < n/2; i++) {
 		IQ[i]   = 0x8000 << 16 | 0x8000;
